requirements.txt
django==3.0.7
python-dotenv==0.13.0
django-cors-headers==3.3.0
django-filter==2.3.0
djangorestframework==3.11.0
celery==4.4.5
requests==2.23.0
django-simple-captcha==0.5.12
djangorestframework-simplejwt==4.4.0
psycopg2-binary==2.8.5
redis==3.5.3
django-oauth-toolkit==1.3.3
